test: 테스트를 위한 클래스 패키지 분리
