+++ mvc/src/test/java/nextstep/mvc/handlerAdapter/HandlerAdapterRegistryTest.java
@@ -4,21 +4,19 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.handlerMapping.HandlerMapping;
-import nextstep.mvc.view.ModelAndView;
+import samples.TestHandlerMapping;
+import samples.TestHandlerAdapter;
 
 class HandlerAdapterRegistryTest {
 
     @Test
     void getHandlerAdapter() {
         HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
-        handlerAdapterRegistry.addHandlerAdapter(new HandlerAdapterForTest());
+        handlerAdapterRegistry.addHandlerAdapter(new TestHandlerAdapter());
 
-        HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(new HandlerForTest());
+        HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(new TestHandlerMapping());
 
-        assertThat(handlerAdapter).isInstanceOf(HandlerAdapterForTest.class);
+        assertThat(handlerAdapter).isInstanceOf(TestHandlerAdapter.class);
     }
 
     @Test
@@ -26,34 +24,8 @@ class HandlerAdapterRegistryTest {
         HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
 
         assertThatThrownBy(
-            () -> handlerAdapterRegistry.getHandlerAdapter(new HandlerForTest()))
+            () -> handlerAdapterRegistry.getHandlerAdapter(new TestHandlerMapping()))
             .isInstanceOf(RuntimeException.class)
             .hasMessageContaining("해당 핸들러를 처리할 수 있는 핸들러 어댑터를 찾지 못했습니다.");
     }
-
-    class HandlerAdapterForTest implements HandlerAdapter {
-        @Override
-        public boolean supports(Object handler) {
-            return handler instanceof HandlerForTest;
-        }
-
-        @Override
-        public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws
-            Exception {
-            return null;
-        }
-    }
-
-    class HandlerForTest implements HandlerMapping {
-        @Override
-        public void initialize() {
-
-        }
-
-        @Override
-        public Object getHandler(HttpServletRequest request) {
-            return null;
-        }
-    }
-
 }
