+++ mvc/src/test/java/samples/TestHandlerAdapter.java
@@ -1,0 +1,20 @@
+package samples;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.handlerAdapter.HandlerAdapter;
+import nextstep.mvc.view.ModelAndView;
+
+public class TestHandlerAdapter implements HandlerAdapter {
+
+    @Override
+    public boolean supports(Object handler) {
+        return handler instanceof TestHandlerMapping;
+    }
+
+    @Override
+    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws
+        Exception {
+        return null;
+    }
+}
