+++ mvc/src/test/java/samples/TestHandlerAdapterRegistry.java
@@ -1,0 +1,4 @@
+package samples;
+
+public class TestHandlerAdapterRegistry {
+}
