+++ mvc/src/test/java/samples/TestHandlerMapping.java
@@ -1,0 +1,16 @@
+package samples;
+
+import jakarta.servlet.http.HttpServletRequest;
+import nextstep.mvc.handlerMapping.HandlerMapping;
+
+public class TestHandlerMapping implements HandlerMapping {
+    @Override
+    public void initialize() {
+
+    }
+
+    @Override
+    public Object getHandler(HttpServletRequest request) {
+        return null;
+    }
+}
